refactor(app-graphql-playground): split factory abstractions into one-per-file in abstractions/ folder
