+++ packages/app-graphql-playground/src/exports/admin/graphql-playground.ts
@@ -1,3 +1,4 @@
-export {PlaygroundTabRegistry} from "~/features/tabRegistry/index.js";
-export {PlaygroundClientFactory} from "~/features/playgroundClient/index.js";
-export {AuthenticatedPlaygroundClientFactory} from "~/features/playgroundClient/index.js";
+export { PlaygroundClient } from "~/features/playgroundClient/index.js";
+export { PlaygroundTabRegistry } from "~/features/tabRegistry/index.js";
+export { PlaygroundClientFactory } from "~/features/playgroundClient/index.js";
+export { AuthenticatedPlaygroundClientFactory } from "~/features/playgroundClient/index.js";
