+++ packages/app-graphql-playground/src/features/playgroundClient/abstractions/PlaygroundClient.ts
