+++ packages/app-graphql-playground/src/features/playgroundClient/index.ts
@@ -1,6 +1,4 @@
-export { PlaygroundClient } from "./abstractions.js";
-export type { IPlaygroundClient } from "./abstractions.js";
-export { PlaygroundClientImpl } from "./PlaygroundClient.js";
+export { PlaygroundClient } from "./abstractions/PlaygroundClient.js";
 export { AuthenticatedPlaygroundClient } from "./AuthenticatedPlaygroundClient.js";
-export { PlaygroundClientFactory } from "./factories.js";
-export { AuthenticatedPlaygroundClientFactory } from "./factories.js";
+export { PlaygroundClientFactory } from "./abstractions/PlaygroundClientFactory.js";
+export { AuthenticatedPlaygroundClientFactory } from "./abstractions/AuthenticatedPlaygroundClientFactory.js";
